+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 Bug fixes:
 
+  - [wr] properly deconstruct array in foreach
   - [lsp] import unresolved classes refactoring: Ensure only unique names are shown when asking user to select an import candidate
   - [lsp] ensure fully qualified filename is used for generate method refactoring #1313
   - [wr] detect branch determination with throw expression